Fix status code bug
