+++ packages/docusaurus-plugin-openapi/src/theme/ObjectTable/styles.module.css
@@ -1,0 +1,3 @@
+.description > div > *:first-child {
+  margin-top: var(--ifm-table-cell-padding);
+}
